Fixes from kernel's checkpatch tool
checkpatch.pl reports many style errors and warnings.  This fixes
all of them except for two instances of >80 character lines.
Fixing those lines would result in less readable code, which is the
litmus test for whether to fix a warning or not.

Part of fixing the errors involved removing the err/warn/info
print macros and using the built in pr_* macros instead.

Change-Id: I0503f72fd3efb4b792d3b4f18eb9596288a3bcb7