Fix delayed inbound message bug
Under heavy load, we are seeing a message from PHYP potentially take
hours (if ever) to be received.  Evidence points to a timing window
where the user thread calling ibmvmc_read sits waiting on a waiter
forever as it missed the wakeup call from the interrupt driven recv_msg.

We weren't checking a conditional between add_wait_queue and schedule.

Change-Id: I4fc63e133da937677bcfd2cda85ef71ac08d957d