Bug fix for buffer check timing
Move the validity check for the userspace data copy ahead of
requesting a buffer from the hmc. This makes it so we won't have
a situation where we request an hmc buffer then fail the userspace
copy check, leaving an unused buffer (since we'll pull a new one
on the next run).

Change-Id: Iacba805a5788c3e2ca0d9f25668d96a7301da622