ibmvmc: Change memory barriers to lese heavyweight
PowerPc wmb() rmb() mb() barriers are very heavyweight,
the new dma_*() calls were specifically made to enforce
ordering of read and writes when accessing coherent
memory that is shared between the CPU and device.

These new calls do not guarantee ordering between
coherent and incoherent memories, but we dont need
that in the usage here.

Signed-off-by: Bryant G. Ly <bryantly@linux.vnet.ibm.com>